+++ plugins/org.polarsys.capella.scenario.editor.dsl.ui/META-INF/MANIFEST.MF
@@ -17,6 +17,7 @@ Require-Bundle: org.polarsys.capella.scenario.editor.dsl,
  org.eclipse.compare,
  org.eclipse.xtext.builder,
  org.polarsys.capella.scenario.editor,
+ org.polarsys.capella.core.model.helpers,
  org.eclipse.xtext.xbase.lib;bundle-version="2.14.0",
  org.eclipse.xtend.lib;bundle-version="2.14.0";resolution:=optional
 Import-Package: org.apache.log4j
